test display with redux store connected
